CRUD_Operations/requirements.txt
colour==0.1.5
mysql-connector-python==8.0.29
Pillow==9.3.0
protobuf==3.20.1
PyMySQL==1.0.2
tkmacosx==1.0.5
